moved waypoints code from plan (back) into the road_map
