+++ smarts/core/road_map.py
@@ -21,15 +21,17 @@
 # to allow for typing to refer to class being defined (RoadMap)
 from __future__ import annotations
 
+from dataclasses import dataclass
 from typing import NamedTuple, List, Sequence, Tuple
 
 import numpy as np
 from shapely.geometry import Polygon
 
-from .coordinates import BoundingBox, Point, Pose, RefLinePoint
+from .coordinates import BoundingBox, Heading, Point, Pose, RefLinePoint
 from .utils.math import (
     fast_quaternion_from_angle,
     min_angles_difference_signed,
+    signed_dist_to_line,
     vec_to_radians,
 )
 
@@ -95,6 +97,19 @@ class RoadMap:
     def random_route(self, max_route_len: int = 10) -> RoadMap.Route:
         raise NotImplementedError()
 
+    def waypoint_paths(
+        self,
+        pose: Pose,
+        lookahead: int,
+        within_radius: float = 5,
+        route: RoadMap.Route = None,
+    ) -> List[List[Waypoint]]:
+        """Computes equally-spaced Waypoints for all lane paths
+        up to lookahead waypoints ahead, starting on the Road containing
+        the nearest Lane aligned with the vehicle's pose within within_radius meters.
+        Constrains paths to the supplied route if specified."""
+        raise NotImplementedError()
+
     class Lane:
         @property
         def lane_id(self) -> str:
@@ -155,6 +170,22 @@ class RoadMap:
             If right-of-way rules are known, these will be in order of priority,
             otherwise they will be in random order.  The boolean flag
             that is returned will be True if the ordering is by priority."""
+            raise NotImplementedError()
+
+        def waypoint_paths_at_point(
+            self, pose: Pose, lookahead: int, route: RoadMap.Route = None
+        ) -> List[List[Waypoint]]:
+            """Computes equally-spaced Waypoints for all lane paths
+            up to lookahead waypoints ahead, starting in this lane at pose.
+            Constrains paths to the supplied route if specified."""
+            raise NotImplementedError()
+
+        def waypoint_paths_at_offset(
+            self, offset: float, lookahead: int = 30, route: RoadMap.Route = None
+        ) -> List[List[Waypoint]]:
+            """Computes equally-spaced Waypoints for all lane paths
+            up to lookahead waypoints ahead, starting offset into this lane.
+            Constrains paths to the supplied route if specified."""
             raise NotImplementedError()
 
         @property
@@ -316,9 +347,6 @@ class RoadMap:
         def geometry(self) -> List[Point]:
             raise NotImplementedError()
 
-    def create_plan_for_mission(self, mission, find_route: bool = True):  # -> Plan
-        raise NotImplementedError()
-
     class Route:
         @property
         def roads(self) -> List[RoadMap.Road]:
@@ -341,3 +369,69 @@ class RoadMap:
         def distance_between(self, start: Point, end: Point) -> float:
             """ Distance along route between two points.  """
             raise NotImplementedError()
+
+
+@dataclass(frozen=True)
+class Waypoint:
+    """Dynamic, based on map and vehicle.  Waypoints
+    start abreast of a vehicle's present location in the nearest Lane
+    and are then interpolated such that they're evenly spaced.
+    These are returned through a vehicle's sensors."""
+
+    # XXX: consider renaming lane_id, lane_index, lane_width
+    #      to nearest_lane_id, nearest_lane_index, nearest_lane_width
+    pos: np.ndarray  # Point positioned on center of lane
+    heading: Heading  # Heading angle of lane at this point (radians)
+    lane_id: str  # ID of lane under waypoint
+    lane_width: float  # Width of lane at this point (meters)
+    speed_limit: float  # Lane speed in m/s
+    lane_index: int  # Index of the lane this waypoint is over. 0 is the outer(right) most lane
+
+    def __eq__(self, other) -> bool:
+        if not isinstance(other, Waypoint):
+            return False
+        return (
+            (self.pos == other.pos).all()
+            and self.heading == other.heading
+            and self.lane_width == other.lane_width
+            and self.speed_limit == other.speed_limit
+            and self.lane_id == other.lane_id
+            and self.lane_index == other.lane_index
+        )
+
+    def __hash__(self):
+        return hash(
+            (
+                *self.pos,
+                self.heading,
+                self.lane_width,
+                self.speed_limit,
+                self.lane_id,
+                self.lane_index,
+            )
+        )
+
+    def relative_heading(self, h: Heading) -> Heading:
+        """Computes relative heading between the given angle and the waypoint heading
+
+        Returns:
+            relative_heading: [-pi..pi]
+        """
+        assert isinstance(
+            h, Heading
+        ), "Heading h ({}) must be an instance of smarts.core.coordinates.Heading".format(
+            type(h)
+        )
+        return self.heading.relative_to(h)
+
+    def signed_lateral_error(self, p) -> float:
+        """Returns the signed lateral distance from the given point to the
+        line formed by the waypoint position and the waypoint heading.
+
+        Negative signals right of line and Positive left of line.
+        """
+        return signed_dist_to_line(p, self.pos, self.heading.direction_vector())
+
+    def dist_to(self, p) -> float:
+        """Calculates straight line distance to the given 2D point"""
+        return np.linalg.norm(self.pos - p[: len(self.pos)])
