+++ smarts/core/sumo_road_network.py
@@ -37,10 +37,16 @@ from shapely.geometry.base import CAP_STYLE, JOIN_STYLE
 from shapely.ops import snap, triangulate
 from trimesh.exchange import gltf
 
-from .coordinates import BoundingBox, Point, Pose, RefLinePoint
-from .road_map import RoadMap
-from .sumo_lanepoints import SumoLanePoints
-from .utils.math import fast_quaternion_from_angle, vec_to_radians
+from .coordinates import BoundingBox, Heading, Point, Pose, RefLinePoint
+from .road_map import RoadMap, Waypoint
+from .sumo_lanepoints import LinkedLanePoint, SumoLanePoints
+from .utils.math import (
+    fast_quaternion_from_angle,
+    inplace_unwrap,
+    radians_to_vec,
+    vec_to_radians,
+    vec_2d,
+)
 
 from smarts.core.utils.sumo import sumolib  # isort:skip
 from sumolib.net.edge import Edge  # isort:skip
@@ -95,12 +101,12 @@ class SumoRoadNetwork(RoadMap):
         )
         self._lanes = {}
         self._roads = {}
-        self.lanepoint_spacing = lanepoint_spacing
-        self.lanepoints = None
+        self._waypoints_cache = SumoRoadNetwork._WaypointsCache()
+        self._lanepoints = None
         if lanepoint_spacing is not None:
             assert lanepoint_spacing > 0
             # XXX: this should be last here since SumoLanePoints() calls road_network methods immediately
-            self.lanepoints = SumoLanePoints(self, spacing=lanepoint_spacing)
+            self._lanepoints = SumoLanePoints(self, spacing=lanepoint_spacing)
 
     @staticmethod
     def _check_net_origin(bbox):
@@ -310,6 +316,40 @@ class SumoRoadNetwork(RoadMap):
             ]
             return (list(set(result)), False)
 
+        def waypoint_paths_at_point(
+            self, pose: Pose, lookahead: int, route: RoadMap.Route = None
+        ) -> List[List[Waypoint]]:
+            road_ids = [road.road_id for road in route.roads] if route else None
+            return self._waypoint_paths_at(pose.position, lookahead, road_ids)
+
+        def waypoint_paths_on_offset(
+            self, offset: float, lookahead: int = 30, route: RoadMap.Route = None
+        ) -> List[List[Waypoint]]:
+            wp_start = self.from_lane_coord(RefLinePoint(offset))
+            road_ids = [road.road_id for road in route.roads] if route else None
+            return self._waypoint_paths_at(wp_start, lookahead, road_ids)
+
+        def _waypoint_paths_at(
+            self,
+            point: Sequence,
+            lookahead: int,
+            filter_road_ids: Sequence[str] = None,
+        ) -> List[List[Waypoint]]:
+            """computes equally-spaced Waypoints for all lane paths
+            up to lookahead waypoints ahead, constrained to filter_road_ids if specified,
+            starting at the nearest LanePoint to point within lane lane_id."""
+            closest_linked_lp = (
+                self._map._lanepoints.closest_linked_lanepoint_on_lane_to_point(
+                    point, self._lane_id
+                )
+            )
+            return self._map._waypoints_starting_at_lanepoint(
+                closest_linked_lp,
+                lookahead,
+                tuple(filter_road_ids) if filter_road_ids else (),
+                tuple(point),
+            )
+
         def buffered_shape(self, width: float = 1.0) -> Polygon:
             return SumoRoadNetwork._buffered_shape(self._sumo_lane.getShape(), width)
 
@@ -516,35 +556,6 @@ class SumoRoadNetwork(RoadMap):
                 return nl.road
         return None
 
-    def create_plan_for_mission(self, mission, find_route: bool = True):  # -> Plan
-        from .sumo_plan import SumoPlan
-
-        return SumoPlan(self, mission, find_route)
-
-    class Route(RoadMap.Route):
-        def __init__(self):
-            self._roads = []
-            self._length = 0
-
-        @property
-        def roads(self) -> List[RoadMap.Road]:
-            return self._roads
-
-        @property
-        def road_length(self) -> float:
-            return self._length
-
-        def add_road(self, road: RoadMap.Road):
-            self._length += road.length
-            self._roads.append(road)
-
-        @cached_property
-        def geometry(self) -> Sequence[Sequence[Tuple[float, float]]]:
-            return [
-                road.buffered_shape(sum([lane._width for lane in road.lanes]))
-                for road in self.roads
-            ]
-
     def generate_routes(
         self,
         start_road: RoadMap.Road,
@@ -643,6 +654,99 @@ class SumoRoadNetwork(RoadMap):
             route.add_road(self.road_by_id(cur_edge.getID()))
             next_edges = list(cur_edge.getOutgoing().keys())
         return route
+
+    def waypoint_paths(
+        self,
+        pose: Pose,
+        lookahead: int,
+        within_radius: float = 5,
+        route: RoadMap.Route = None,
+    ) -> List[List[Waypoint]]:
+        if route:
+            if route.roads:
+                road_ids = [road.road_id for road in route.roads]
+            else:
+                road_ids = self._resolve_in_junction(pose)
+            if road_ids:
+                return self._waypoint_paths_along_route(
+                    pose.position, lookahead, road_ids
+                )
+        closest_lps = self._lanepoints.closest_lanepoints(
+            [pose], within_radius=within_radius
+        )
+        closest_lane = closest_lps[0].lane
+        # TAI: the above lines could be replaced by:
+        # closest_lane = self.nearest_lane(pose.point, radius=within_radius)
+        waypoint_paths = []
+        for lane in closest_lane.road.lanes:
+            waypoint_paths += lane._waypoint_paths_at(pose.position, lookahead)
+        return sorted(waypoint_paths, key=lambda p: p[0].lane_index)
+
+    def _resolve_in_junction(self, pose: Pose) -> List[str]:
+        # This is so that the waypoints don't jump between connections
+        # when we don't know which lane we're on in a junction.
+        # We take the 10 closest lanepoints then filter down to that which has
+        # the closest heading. This way we get the lanepoint on our lane instead of
+        # a potentially closer lane that is on a different junction connection.
+        closest_lps = self._lanepoints.closest_lanepoints([pose], within_radius=None)
+        closest_lps.sort(key=lambda lp: abs(pose.heading - lp.pose.heading))
+        lane = closest_lps[0].lane
+        if not lane.in_junction:
+            return []
+        road_ids = [lane.road.road_id]
+        next_roads = lane.road.outgoing_roads
+        assert (
+            len(next_roads) <= 1
+        ), "A junction is expected to have <= 1 outgoing roads"
+        if next_roads:
+            road_ids.append(next_roads[0].road_id)
+        return road_ids
+
+    def _waypoint_paths_along_route(
+        self, point, lookahead: int, route: Sequence[str]
+    ) -> List[List[Waypoint]]:
+        """finds the closest lane to vehicle's position that is on its route,
+        then gets waypoint paths from all lanes in its edge there."""
+        assert len(route) > 0, f"Expected at least 1 road in the route, got: {route}"
+        closest_llp_on_each_route_road = [
+            self._lanepoints.closest_linked_lanepoint_on_road(point, road)
+            for road in route
+        ]
+        closest_linked_lp = min(
+            closest_llp_on_each_route_road,
+            key=lambda l_lp: np.linalg.norm(
+                vec_2d(l_lp.lp.pose.position) - vec_2d(point)
+            ),
+        )
+        closest_lane = closest_linked_lp.lp.lane
+        waypoint_paths = []
+        for lane in closest_lane.road.lanes:
+            waypoint_paths += lane._waypoint_paths_at(point, lookahead, route)
+        return sorted(waypoint_paths, key=lambda p: p[0].lane_index)
+
+    class Route(RoadMap.Route):
+        def __init__(self):
+            self._roads = []
+            self._length = 0
+
+        @property
+        def roads(self) -> List[RoadMap.Road]:
+            return self._roads
+
+        @property
+        def road_length(self) -> float:
+            return self._length
+
+        def add_road(self, road: RoadMap.Road):
+            self._length += road.length
+            self._roads.append(road)
+
+        @cached_property
+        def geometry(self) -> Sequence[Sequence[Tuple[float, float]]]:
+            return [
+                road.buffered_shape(sum([lane._width for lane in road.lanes]))
+                for road in self.roads
+            ]
 
     def _compute_road_polygons(self):
         lane_to_poly = {}
@@ -895,3 +999,204 @@ class SumoRoadNetwork(RoadMap):
         connection_lane = self._graph.getLane(connection_lane_id)
 
         return connection_lane.getEdge().getID()
+
+    class _WaypointsCache:
+        def __init__(self):
+            self.lookahead = 0
+            self.point = (0, 0, 0)
+            self.filter_road_ids = ()
+            self._starts = {}
+
+        def _match(self, lookahead, point, filter_road_ids) -> bool:
+            return (
+                lookahead <= self.lookahead
+                and point[0] == self.point[0]
+                and point[1] == self.point[1]
+                and filter_road_ids == self.filter_road_ids
+            )
+
+        def update(
+            self,
+            lookahead: int,
+            point: Tuple[float, float, float],
+            filter_road_ids: tuple,
+            llp: LinkedLanePoint,
+            paths: List[List[Waypoint]],
+        ):
+            if not self._match(lookahead, point, filter_road_ids):
+                self.lookahead = lookahead
+                self.point = point
+                self.filter_road_ids = filter_road_ids
+                self._starts = {}
+            self._starts[llp.lp.lane.index] = paths
+
+        def query(
+            self,
+            lookahead: int,
+            point: Tuple[float, float, float],
+            filter_road_ids: tuple,
+            llp: LinkedLanePoint,
+        ) -> List[List[Waypoint]]:
+            if self._match(lookahead, point, filter_road_ids):
+                hit = self._starts.get(llp.lp.lane.index, None)
+                if hit:
+                    # consider just returning all of them (not slicing)?
+                    return [path[: (lookahead + 1)] for path in hit]
+            return None
+
+    def _waypoints_starting_at_lanepoint(
+        self,
+        lanepoint: LinkedLanePoint,
+        lookahead: int,
+        filter_road_ids: tuple,
+        point: Tuple[float, float, float],
+    ) -> List[List[Waypoint]]:
+        """computes equally-spaced Waypoints for all lane paths starting at lanepoint
+        up to lookahead waypoints ahead, constrained to filter_road_ids if specified."""
+
+        # The following acts sort of like lru_cache(1), but it allows
+        # for lookahead to be <= to the cached value...
+        cache_paths = self._waypoints_cache.query(
+            lookahead, point, filter_road_ids, lanepoint
+        )
+        if cache_paths:
+            return cache_paths
+
+        lanepoint_paths = self._lanepoints.paths_starting_at_lanepoint(
+            lanepoint, lookahead, filter_road_ids
+        )
+        result = [
+            SumoRoadNetwork._equally_spaced_path(path, point)
+            for path in lanepoint_paths
+        ]
+
+        self._waypoints_cache.update(
+            lookahead, point, filter_road_ids, lanepoint, result
+        )
+
+        return result
+
+    @staticmethod
+    def _equally_spaced_path(
+        path: Sequence[LinkedLanePoint], point: Tuple[float, float, float]
+    ) -> List[Waypoint]:
+        """given a list of LanePoints starting near point, that may not be evenly spaced,
+        returns the same number of Waypoints that are evenly spaced and start at point."""
+
+        continuous_variables = [
+            "positions_x",
+            "positions_y",
+            "headings",
+            "lane_width",
+            "speed_limit",
+        ]
+        discrete_variables = ["lane_id", "lane_index"]
+
+        ref_lanepoints_coordinates = {
+            parameter: [] for parameter in (continuous_variables + discrete_variables)
+        }
+        for idx, lanepoint in enumerate(path):
+            if lanepoint.is_inferred and 0 < idx < len(path) - 1:
+                continue
+            ref_lanepoints_coordinates["positions_x"].append(
+                lanepoint.lp.pose.position[0]
+            )
+            ref_lanepoints_coordinates["positions_y"].append(
+                lanepoint.lp.pose.position[1]
+            )
+            ref_lanepoints_coordinates["headings"].append(
+                lanepoint.lp.pose.heading.as_bullet
+            )
+            ref_lanepoints_coordinates["lane_id"].append(lanepoint.lp.lane.lane_id)
+            ref_lanepoints_coordinates["lane_index"].append(lanepoint.lp.lane.index)
+            ref_lanepoints_coordinates["lane_width"].append(lanepoint.lp.lane._width)
+            ref_lanepoints_coordinates["speed_limit"].append(
+                lanepoint.lp.lane.speed_limit
+            )
+
+        ref_lanepoints_coordinates["headings"] = inplace_unwrap(
+            ref_lanepoints_coordinates["headings"]
+        )
+        first_lp_heading = ref_lanepoints_coordinates["headings"][0]
+        lp_position = path[0].lp.pose.position[:2]
+        vehicle_pos = np.array(point[:2])
+        heading_vec = np.array(radians_to_vec(first_lp_heading))
+        projected_distant_lp_vehicle = np.inner(
+            (vehicle_pos - lp_position), heading_vec
+        )
+
+        ref_lanepoints_coordinates["positions_x"][0] = (
+            lp_position[0] + projected_distant_lp_vehicle * heading_vec[0]
+        )
+        ref_lanepoints_coordinates["positions_y"][0] = (
+            lp_position[1] + projected_distant_lp_vehicle * heading_vec[1]
+        )
+        # To ensure that the distance between waypoints are equal, we used
+        # interpolation approach inspired by:
+        # https://stackoverflow.com/a/51515357
+        cumulative_path_dist = np.cumsum(
+            np.sqrt(
+                np.ediff1d(ref_lanepoints_coordinates["positions_x"], to_begin=0) ** 2
+                + np.ediff1d(ref_lanepoints_coordinates["positions_y"], to_begin=0) ** 2
+            )
+        )
+
+        if len(cumulative_path_dist) <= 1:
+            lp = path[0].lp
+            return [
+                Waypoint(
+                    pos=lp.pose.position,
+                    heading=lp.pose.heading,
+                    lane_width=lp.lane._width,
+                    speed_limit=lp.lane.speed_limit,
+                    lane_id=lp.lane.lane_id,
+                    lane_index=lp.lane.index,
+                )
+            ]
+
+        evenly_spaced_cumulative_path_dist = np.linspace(
+            0, cumulative_path_dist[-1], len(path)
+        )
+
+        evenly_spaced_coordinates = {}
+        for variable in continuous_variables:
+            evenly_spaced_coordinates[variable] = np.interp(
+                evenly_spaced_cumulative_path_dist,
+                cumulative_path_dist,
+                ref_lanepoints_coordinates[variable],
+            )
+
+        for variable in discrete_variables:
+            ref_coordinates = ref_lanepoints_coordinates[variable]
+            evenly_spaced_coordinates[variable] = []
+            jdx = 0
+            for idx in range(len(path)):
+                while (
+                    jdx + 1 < len(cumulative_path_dist)
+                    and evenly_spaced_cumulative_path_dist[idx]
+                    > cumulative_path_dist[jdx + 1]
+                ):
+                    jdx += 1
+
+                evenly_spaced_coordinates[variable].append(ref_coordinates[jdx])
+            evenly_spaced_coordinates[variable].append(ref_coordinates[-1])
+
+        equally_spaced_path = []
+        for idx in range(len(path)):
+            equally_spaced_path.append(
+                Waypoint(
+                    pos=np.array(
+                        [
+                            evenly_spaced_coordinates["positions_x"][idx],
+                            evenly_spaced_coordinates["positions_y"][idx],
+                        ]
+                    ),
+                    heading=Heading(evenly_spaced_coordinates["headings"][idx]),
+                    lane_width=evenly_spaced_coordinates["lane_width"][idx],
+                    speed_limit=evenly_spaced_coordinates["speed_limit"][idx],
+                    lane_id=evenly_spaced_coordinates["lane_id"][idx],
+                    lane_index=evenly_spaced_coordinates["lane_index"][idx],
+                )
+            )
+
+        return equally_spaced_path
